+++ src/main/kotlin/icu/windea/ut/toolbox/jsonSchema/LanguageSettings.kt
@@ -1,7 +1,0 @@
-package icu.windea.ut.toolbox.jsonSchema
-
-data class LanguageSettings(
-    val declarations: Set<String>,
-    val references: Set<String>,
-    val completion: Set<String>,
-)